create block script directory; refactor block.cs
I refactored most of the block code to give it a state
machine layout. I think this might be easier to maintain in the
future, and also easier for each other to digest in the future.
It also will follow a similar development to the Player.cs class.
A few optimizations were made as well.